Fix a bug; add more tests (including regression tests)
