+++ projects/batfish/src/test/java/org/batfish/dataplane/ibdp/IncrementalDataPlanePluginTest.java
@@ -3,11 +3,9 @@ package org.batfish.dataplane.ibdp;
 import static org.batfish.datamodel.Configuration.DEFAULT_VRF_NAME;
 import static org.batfish.datamodel.matchers.AbstractRouteMatchers.hasPrefix;
 import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.Matchers.greaterThan;
 import static org.hamcrest.Matchers.hasItem;
 import static org.hamcrest.Matchers.hasSize;
 import static org.hamcrest.Matchers.in;
-import static org.hamcrest.Matchers.lessThan;
 import static org.hamcrest.Matchers.not;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertThat;
@@ -19,12 +17,9 @@ import com.google.common.collect.ImmutableSortedMap;
 import com.google.common.collect.ImmutableSortedSet;
 import com.google.common.graph.ValueGraph;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Set;
 import java.util.SortedMap;
@@ -36,13 +31,10 @@ import org.batfish.common.BatfishLogger;
 import org.batfish.common.plugin.DataPlanePlugin;
 import org.batfish.common.plugin.DataPlanePlugin.ComputeDataPlaneResult;
 import org.batfish.datamodel.AbstractRoute;
-import org.batfish.datamodel.AsPath;
 import org.batfish.datamodel.BgpActivePeerConfig;
 import org.batfish.datamodel.BgpPeerConfigId;
 import org.batfish.datamodel.BgpProcess;
-import org.batfish.datamodel.BgpRoute;
 import org.batfish.datamodel.BgpSessionProperties;
-import org.batfish.datamodel.BgpTieBreaker;
 import org.batfish.datamodel.Configuration;
 import org.batfish.datamodel.ConfigurationFormat;
 import org.batfish.datamodel.DataPlane;
@@ -51,24 +43,18 @@ import org.batfish.datamodel.GeneratedRoute.Builder;
 import org.batfish.datamodel.Interface;
 import org.batfish.datamodel.InterfaceAddress;
 import org.batfish.datamodel.Ip;
-import org.batfish.datamodel.MultipathEquivalentAsPathMatchMode;
 import org.batfish.datamodel.NetworkFactory;
-import org.batfish.datamodel.OriginType;
 import org.batfish.datamodel.Prefix;
 import org.batfish.datamodel.Route;
-import org.batfish.datamodel.RoutingProtocol;
 import org.batfish.datamodel.StaticRoute;
 import org.batfish.datamodel.Topology;
 import org.batfish.datamodel.Vrf;
 import org.batfish.datamodel.collections.RoutesByVrf;
 import org.batfish.datamodel.routing_policy.RoutingPolicy;
 import org.batfish.datamodel.routing_policy.statement.SetDefaultPolicy;
-import org.batfish.dataplane.rib.BgpBestPathRib;
-import org.batfish.dataplane.rib.BgpMultipathRib;
 import org.batfish.main.Batfish;
 import org.batfish.main.BatfishTestUtils;
 import org.batfish.main.TestrigText;
-import org.hamcrest.Matcher;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -178,212 +164,6 @@ public class IncrementalDataPlanePluginTest {
     return ImmutableSortedMap.of(CORE_NAME, core, "n1", n1, "n2", n2);
   }
 
-  private void testBgpAsPathMultipathHelper(
-      MultipathEquivalentAsPathMatchMode multipathEquivalentAsPathMatchMode,
-      boolean primeBestPathInMultipathBgpRib,
-      boolean expectRoute2,
-      boolean expectRoute3a,
-      boolean expectRoute3b,
-      boolean expectRoute3c,
-      boolean expectRoute3d) {
-    /*
-     * Properties of the routes
-     */
-    // Should appear only for path-length match
-    AsPath asPath2 = AsPath.ofSingletonAsSets(2L, 4L, 6L);
-    // Should appear only for first-as match and path-length match
-    AsPath asPath3a = AsPath.ofSingletonAsSets(3L, 5L, 6L);
-    // Should never appear
-    AsPath asPath3b = AsPath.ofSingletonAsSets(3L, 4L, 4L, 6L);
-    // Should always appear
-    AsPath bestAsPath = AsPath.ofSingletonAsSets(3L, 4L, 6L);
-    AsPath asPath3c = bestAsPath;
-    AsPath asPath3d = bestAsPath;
-    Ip nextHop2 = new Ip("2.0.0.0");
-    Ip nextHop3a = new Ip("3.0.0.1");
-    Ip nextHop3b = new Ip("3.0.0.2");
-    Ip nextHop3c = new Ip("3.0.0.3");
-    Ip nextHop3d = new Ip("3.0.0.4");
-
-    /*
-     * Common attributes for all routes
-     */
-    Prefix p = Prefix.ZERO;
-    BgpRoute.Builder b =
-        new BgpRoute.Builder()
-            .setNetwork(p)
-            .setProtocol(RoutingProtocol.BGP)
-            .setOriginType(OriginType.INCOMPLETE);
-
-    /*
-     * Boilerplate virtual-router setup
-     */
-    String hostname = "r1";
-    Configuration c =
-        BatfishTestUtils.createTestConfiguration(hostname, ConfigurationFormat.CISCO_IOS);
-    BgpProcess proc = new BgpProcess();
-    c.getVrfs().computeIfAbsent(DEFAULT_VRF_NAME, Vrf::new).setBgpProcess(proc);
-
-    /*
-     * Instantiate routes
-     */
-    BgpRoute route2 =
-        b.setAsPath(asPath2)
-            .setNextHopIp(nextHop2)
-            .setOriginatorIp(nextHop2)
-            .setReceivedFromIp(nextHop2)
-            .build();
-    BgpRoute route3a =
-        b.setAsPath(asPath3a)
-            .setNextHopIp(nextHop3a)
-            .setOriginatorIp(nextHop3a)
-            .setReceivedFromIp(nextHop3a)
-            .build();
-    BgpRoute route3b =
-        b.setAsPath(asPath3b)
-            .setNextHopIp(nextHop3b)
-            .setOriginatorIp(nextHop3b)
-            .setReceivedFromIp(nextHop3b)
-            .build();
-    BgpRoute route3c =
-        b.setAsPath(asPath3c)
-            .setNextHopIp(nextHop3c)
-            .setOriginatorIp(nextHop3c)
-            .setReceivedFromIp(nextHop3c)
-            .build();
-    BgpRoute route3d =
-        b.setAsPath(asPath3d)
-            .setNextHopIp(nextHop3d)
-            .setOriginatorIp(nextHop3d)
-            .setReceivedFromIp(nextHop3d)
-            .build();
-
-    /*
-     * Set the as-path match mode prior to instantiating bgp multipath RIB
-     */
-    proc.setMultipathEquivalentAsPathMatchMode(multipathEquivalentAsPathMatchMode);
-    BgpMultipathRib bmr = new BgpMultipathRib(proc.getMultipathEquivalentAsPathMatchMode());
-
-    /*
-     * Prime bgp multipath RIB with best path for the prefix
-     */
-    if (primeBestPathInMultipathBgpRib) {
-      bmr.setBestAsPaths(Collections.singletonMap(p, bestAsPath));
-    }
-
-    /*
-     * Add routes to multipath RIB.
-     */
-    bmr.mergeRoute(route2);
-    bmr.mergeRoute(route3a);
-    bmr.mergeRoute(route3b);
-    bmr.mergeRoute(route3c);
-    bmr.mergeRoute(route3d);
-
-    /*
-     * Initialize the matchers with respect to the output route set
-     */
-    Set<BgpRoute> postMergeRoutes = bmr.getRoutes();
-    Matcher<BgpRoute> present = in(postMergeRoutes);
-    Matcher<BgpRoute> absent = not(present);
-
-    /*
-     * ASSERTIONS:
-     * Only the expected routes for the given match mode should be present at end
-     */
-    assertThat(route2, expectRoute2 ? present : absent);
-    assertThat(route3a, expectRoute3a ? present : absent);
-    assertThat(route3b, expectRoute3b ? present : absent);
-    assertThat(route3c, expectRoute3c ? present : absent);
-    assertThat(route3c, expectRoute3d ? present : absent);
-  }
-
-  @Test
-  public void testBgpAsPathMultipathExactPath() {
-    /*
-     * Only routes with first-as matching that of best as path should appear in RIB post-merge.
-     */
-    testBgpAsPathMultipathHelper(
-        MultipathEquivalentAsPathMatchMode.EXACT_PATH, true, false, false, false, true, true);
-  }
-
-  @Test
-  public void testBgpAsPathMultipathFirstAs() {
-    /*
-     * Only routes with first-as matching that of best as path should appear in RIB post-merge.
-     */
-    testBgpAsPathMultipathHelper(
-        MultipathEquivalentAsPathMatchMode.FIRST_AS, true, false, true, false, true, true);
-  }
-
-  @Test
-  public void testBgpAsPathMultipathPathLength() {
-    /*
-     * All routes with same as-path-length as that of best as-path should appear in RIB post-merge.
-     */
-    testBgpAsPathMultipathHelper(
-        MultipathEquivalentAsPathMatchMode.PATH_LENGTH, true, true, true, false, true, true);
-  }
-
-  @Test
-  public void testBgpAsPathMultipathUnprimed() {
-    /*
-     * Without priming best as path map, all paths except the longer one should be considered
-     * equivalent. Results should be independent of chosen mode.
-     */
-    for (MultipathEquivalentAsPathMatchMode mode : MultipathEquivalentAsPathMatchMode.values()) {
-      testBgpAsPathMultipathHelper(mode, false, true, true, false, true, true);
-    }
-  }
-
-  @Test
-  public void testBgpCompareOriginId() {
-    String hostname = "r1";
-    Configuration c =
-        BatfishTestUtils.createTestConfiguration(hostname, ConfigurationFormat.CISCO_IOS);
-    BgpProcess proc = new BgpProcess();
-    c.getVrfs().computeIfAbsent(DEFAULT_VRF_NAME, Vrf::new).setBgpProcess(proc);
-    BgpBestPathRib bbr = BgpBestPathRib.initial(null, null);
-    BgpMultipathRib bmr = new BgpMultipathRib(proc.getMultipathEquivalentAsPathMatchMode());
-    Prefix p = Prefix.ZERO;
-    BgpRoute.Builder b = new BgpRoute.Builder().setNetwork(p).setProtocol(RoutingProtocol.IBGP);
-
-    /*
-     *  Initialize with different originator ips, which should not affect comparison of routes with
-     *  different origin type.
-     */
-    Map<OriginType, List<BgpRoute>> routesByOriginType = new LinkedHashMap<>();
-    for (OriginType originType : OriginType.values()) {
-      List<BgpRoute> routes =
-          routesByOriginType.computeIfAbsent(originType, o -> new ArrayList<>());
-      routes.add(
-          b.setOriginatorIp(Ip.ZERO).setReceivedFromIp(Ip.ZERO).setOriginType(originType).build());
-      routes.add(
-          b.setOriginatorIp(Ip.MAX).setReceivedFromIp(Ip.MAX).setOriginType(originType).build());
-    }
-
-    /*
-     * Whenever origin type is different, it should be overriding factor in preference.
-     */
-    for (OriginType o1 : OriginType.values()) {
-      List<BgpRoute> lhsList = routesByOriginType.get(o1);
-      for (OriginType o2 : OriginType.values()) {
-        List<BgpRoute> rhsList = routesByOriginType.get(o2);
-        for (BgpRoute lhs : lhsList) {
-          for (BgpRoute rhs : rhsList) {
-            if (o1.getPreference() > o2.getPreference()) {
-              assertThat(bbr.comparePreference(lhs, rhs), greaterThan(0));
-              assertThat(bmr.comparePreference(lhs, rhs), greaterThan(0));
-            } else if (o1.getPreference() < o2.getPreference()) {
-              assertThat(bbr.comparePreference(lhs, rhs), lessThan(0));
-              assertThat(bmr.comparePreference(lhs, rhs), lessThan(0));
-            }
-          }
-        }
-      }
-    }
-  }
-
   @Test
   public void testBgpOscillation() throws IOException {
     String testrigName = "bgp-oscillation";
@@ -400,86 +180,6 @@ public class IncrementalDataPlanePluginTest {
 
     // Really just test that no exception is thrown
     dataPlanePlugin.computeDataPlane(false);
-  }
-
-  @Test
-  public void testBgpTieBreaker() {
-    String hostname = "r1";
-    Configuration c =
-        BatfishTestUtils.createTestConfiguration(hostname, ConfigurationFormat.CISCO_IOS);
-    BgpProcess proc = new BgpProcess();
-    c.getVrfs().computeIfAbsent(DEFAULT_VRF_NAME, Vrf::new).setBgpProcess(proc);
-
-    // good for both ebgp and ibgp
-    BgpMultipathRib bmr = new BgpMultipathRib(proc.getMultipathEquivalentAsPathMatchMode());
-    // ebgp
-    BgpBestPathRib ebgpBpr = BgpBestPathRib.initial(null, null);
-    BgpRoute.Builder ebgpBuilder =
-        new BgpRoute.Builder()
-            .setNetwork(Prefix.ZERO)
-            .setOriginType(OriginType.INCOMPLETE)
-            .setOriginatorIp(Ip.ZERO)
-            .setProtocol(RoutingProtocol.BGP)
-            .setReceivedFromIp(Ip.ZERO);
-    BgpRoute ebgpOlderHigherOriginator =
-        ebgpBuilder.setOriginatorIp(Ip.MAX).setReceivedFromIp(new Ip("1.1.1.1")).build();
-    BgpRoute ebgpNewerHigherOriginator =
-        ebgpBuilder.setOriginatorIp(Ip.MAX).setReceivedFromIp(new Ip("1.1.1.2")).build();
-    BgpRoute ebgpLowerOriginator = ebgpBuilder.setOriginatorIp(Ip.ZERO).build();
-    // ibgp
-    BgpBestPathRib ibgpBpr = BgpBestPathRib.initial(null, null);
-    BgpRoute.Builder ibgpBuilder =
-        new BgpRoute.Builder()
-            .setNetwork(Prefix.ZERO)
-            .setOriginType(OriginType.INCOMPLETE)
-            .setOriginatorIp(Ip.ZERO)
-            .setProtocol(RoutingProtocol.IBGP)
-            .setReceivedFromIp(Ip.ZERO);
-    BgpRoute ibgpOlderHigherOriginator =
-        ibgpBuilder.setOriginatorIp(Ip.MAX).setReceivedFromIp(new Ip("1.1.1.1")).build();
-    BgpRoute ibgpNewerHigherOriginator =
-        ibgpBuilder.setOriginatorIp(Ip.MAX).setReceivedFromIp(new Ip("1.1.1.2")).build();
-    BgpRoute ibgpLowerOriginator = ibgpBuilder.setOriginatorIp(Ip.ZERO).build();
-
-    ebgpBpr.mergeRoute(ebgpOlderHigherOriginator);
-    ibgpBpr.mergeRoute(ibgpOlderHigherOriginator);
-
-    /*
-     * Given default tie-breaking, and all more important attributes being equivalent:
-     * - When comparing two eBGP adverts, best-path rib prefers older advert.
-     * - If neither is older, or one is iBGP, best-path rib prefers advert with higher router-id.
-     * - Multipath RIB ignores both age and router-id, seeing both adverts as equal.
-     */
-
-    // Test age comparisons first
-    assertThat(
-        ebgpBpr.comparePreference(ebgpNewerHigherOriginator, ebgpOlderHigherOriginator),
-        lessThan(0));
-    assertThat(bmr.comparePreference(ebgpNewerHigherOriginator, ebgpLowerOriginator), equalTo(0));
-    assertThat(
-        ibgpBpr.comparePreference(ibgpNewerHigherOriginator, ibgpLowerOriginator), lessThan(0));
-    assertThat(bmr.comparePreference(ibgpNewerHigherOriginator, ibgpLowerOriginator), equalTo(0));
-
-    /*
-     * No two routes have the same arrival time, so force different non-default tie breaker that
-     * will *not* break the tie, and we get to originator comparisons.
-     */
-    proc.setTieBreaker(BgpTieBreaker.CLUSTER_LIST_LENGTH);
-    // Test IP
-    assertThat(
-        ebgpBpr.comparePreference(ebgpNewerHigherOriginator, ebgpLowerOriginator), lessThan(0));
-    assertThat(
-        ibgpBpr.comparePreference(ibgpNewerHigherOriginator, ibgpLowerOriginator), lessThan(0));
-
-    /*
-     * Finally check the neighbor IP is used as last resolution step
-     */
-    assertThat(
-        ebgpBpr.comparePreference(ebgpNewerHigherOriginator, ebgpOlderHigherOriginator),
-        lessThan(0));
-    assertThat(
-        ibgpBpr.comparePreference(ibgpNewerHigherOriginator, ibgpOlderHigherOriginator),
-        lessThan(0));
   }
 
   @Test
@@ -512,75 +212,6 @@ public class IncrementalDataPlanePluginTest {
     assertThat(r3Loopback0Prefix, in(r1Prefixes));
     // Check the other direction (r1loopback is accepted by r3)
     assertThat(r1Loopback0Prefix, in(r3Prefixes));
-  }
-
-  @Test
-  public void testContainsRoute() {
-    String hostname = "r1";
-    Configuration c =
-        BatfishTestUtils.createTestConfiguration(hostname, ConfigurationFormat.CISCO_IOS);
-    BgpProcess proc = new BgpProcess();
-    c.getVrfs().computeIfAbsent(DEFAULT_VRF_NAME, Vrf::new).setBgpProcess(proc);
-    BgpBestPathRib bbr = BgpBestPathRib.initial(null, null);
-    BgpMultipathRib bmr = new BgpMultipathRib(proc.getMultipathEquivalentAsPathMatchMode());
-    Ip ip1 = new Ip("1.0.0.0");
-    Ip ip2 = new Ip("2.2.0.0");
-    BgpRoute.Builder b1 =
-        new BgpRoute.Builder()
-            .setNextHopIp(Ip.ZERO)
-            .setOriginType(OriginType.INCOMPLETE)
-            .setOriginatorIp(Ip.ZERO)
-            .setProtocol(RoutingProtocol.BGP)
-            .setReceivedFromIp(Ip.ZERO);
-    BgpRoute.Builder b2 =
-        new BgpRoute.Builder()
-            .setNextHopIp(Ip.ZERO)
-            .setOriginType(OriginType.INCOMPLETE)
-            .setOriginatorIp(Ip.MAX)
-            .setProtocol(RoutingProtocol.BGP)
-            .setReceivedFromIp(Ip.ZERO);
-
-    /*
-     * Toss a bunch of different routes in each RIB. In the best-path rib, only lower originatorIp
-     * routes should remain. In the multipath RIB, all routes should remain.
-     */
-    for (int i = 8; i <= Prefix.MAX_PREFIX_LENGTH; i++) {
-      Prefix p = new Prefix(ip1, i);
-      b1.setNetwork(p);
-      b2.setNetwork(p);
-      bbr.mergeRoute(b1.build());
-      bbr.mergeRoute(b2.build());
-      bmr.mergeRoute(b1.build());
-      bmr.mergeRoute(b2.build());
-    }
-    for (int i = 16; i <= Prefix.MAX_PREFIX_LENGTH; i++) {
-      Prefix p = new Prefix(ip2, i);
-      b1.setNetwork(p);
-      b2.setNetwork(p);
-      bbr.mergeRoute(b1.build());
-      bbr.mergeRoute(b2.build());
-      bmr.mergeRoute(b1.build());
-      bmr.mergeRoute(b2.build());
-    }
-    for (int i = 8; i <= Prefix.MAX_PREFIX_LENGTH; i++) {
-      Prefix p = new Prefix(ip1, i);
-      assertTrue(bbr.containsRoute(b1.setNetwork(p).build()));
-      b1.setNetwork(p);
-      b2.setNetwork(p);
-      assertTrue(bbr.containsRoute(b1.build()));
-      assertFalse(bbr.containsRoute(b2.build()));
-      assertTrue(bmr.containsRoute(b1.build()));
-      assertTrue(bmr.containsRoute(b2.build()));
-    }
-    for (int i = 16; i <= Prefix.MAX_PREFIX_LENGTH; i++) {
-      Prefix p = new Prefix(ip2, i);
-      b1.setNetwork(p);
-      b2.setNetwork(p);
-      assertTrue(bbr.containsRoute(b1.build()));
-      assertFalse(bbr.containsRoute(b2.build()));
-      assertTrue(bmr.containsRoute(b1.build()));
-      assertTrue(bmr.containsRoute(b2.build()));
-    }
   }
 
   @Test
