+++ projects/batfish/src/test/java/org/batfish/dataplane/rib/BgpRibTest.java
@@ -2,10 +2,21 @@ package org.batfish.dataplane.rib;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
+import static org.hamcrest.Matchers.greaterThan;
 import static org.hamcrest.Matchers.hasSize;
+import static org.hamcrest.Matchers.in;
+import static org.hamcrest.Matchers.lessThan;
+import static org.hamcrest.Matchers.not;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 
 import com.google.common.collect.ImmutableSortedSet;
+import java.util.ArrayList;
 import java.util.Collections;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 import org.batfish.datamodel.AsPath;
 import org.batfish.datamodel.BgpRoute;
 import org.batfish.datamodel.BgpTieBreaker;
@@ -15,6 +26,7 @@ import org.batfish.datamodel.OriginType;
 import org.batfish.datamodel.Prefix;
 import org.batfish.datamodel.RoutingProtocol;
 import org.batfish.datamodel.StaticRoute;
+import org.hamcrest.Matcher;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -28,6 +40,7 @@ public class BgpRibTest {
 
   private BgpRoute.Builder _rb;
   private BgpRib _multiPathRib;
+  private BgpRib _bestPathRib;
 
   @Rule public ExpectedException thrown = ExpectedException.none();
 
@@ -53,6 +66,7 @@ public class BgpRibTest {
             BgpTieBreaker.ROUTER_ID,
             null,
             MultipathEquivalentAsPathMatchMode.EXACT_PATH);
+    _bestPathRib = new BgpRib(null, null, BgpTieBreaker.ROUTER_ID, 1, null);
   }
 
   @Test
@@ -357,5 +371,376 @@ public class BgpRibTest {
         _rb.setNetwork(Prefix.parse("10.1.1.0/24")).setOriginatorIp(new Ip("22.22.22.22")).build());
     assertThat(_multiPathRib.getRoutes(), hasSize(4));
     assertThat(_multiPathRib.getBestPathRoutes(), hasSize(3));
+  }
+
+  @Test
+  public void testBestPathSelectionWithoutTieBreaking() {
+    _bestPathRib.mergeRoute(_rb.build());
+    BgpRoute bestPath = _rb.setLocalPreference(200).build();
+    _bestPathRib.mergeRoute(bestPath);
+
+    assertThat(_bestPathRib.getRoutes(), equalTo(Collections.singleton(bestPath)));
+  }
+
+  @Test
+  public void testBestPathSelectionTieBreakRouterId() {
+    _bestPathRib.mergeRoute(_rb.build());
+    // Lower originator IP is better
+    BgpRoute bestPath = _rb.setOriginatorIp(new Ip("1.1.0.1")).build();
+    _bestPathRib.mergeRoute(bestPath);
+
+    assertThat(_bestPathRib.getRoutes(), equalTo(Collections.singleton(bestPath)));
+  }
+
+  @Test
+  public void testBestPathSelectionTieBreakReceivedFrom() {
+    _bestPathRib.mergeRoute(_rb.build());
+    // Lower IP is better
+    BgpRoute bestPath = _rb.setReceivedFromIp(new Ip("1.1.0.1")).build();
+    _bestPathRib.mergeRoute(bestPath);
+
+    assertThat(_bestPathRib.getRoutes(), equalTo(Collections.singleton(bestPath)));
+  }
+
+  @Test
+  public void testBestPathSelectionTieBreakingArrivalOrder() {
+    _bestPathRib = new BgpRib(null, null, BgpTieBreaker.ARRIVAL_ORDER, 1, null);
+    BgpRoute bestPath = _rb.build();
+    _bestPathRib.mergeRoute(bestPath);
+    // Oldest route should win despite newer having lower in Originator IP
+    _bestPathRib.mergeRoute(_rb.setOriginatorIp(new Ip("1.1.0.1")).build());
+
+    assertThat(_bestPathRib.getRoutes(), equalTo(Collections.singleton(bestPath)));
+  }
+
+  //////////////////////////////////////////////////////////////////////////////////
+
+  // NOTE: Tests below are old and may be somewhat hard to comprehend.
+  // For now, keeping them as regression tests.
+
+  //////////////////////////////////////////////////////////////////////////////////
+
+  private static void testBgpAsPathMultipathHelper(
+      MultipathEquivalentAsPathMatchMode multipathEquivalentAsPathMatchMode,
+      boolean expectRoute2,
+      boolean expectRoute3a,
+      boolean expectRoute3b,
+      boolean expectRoute3c,
+      boolean expectRoute3d) {
+    /*
+     * Properties of the routes
+     */
+    // Should appear only for path-length match
+    AsPath asPath2 = AsPath.ofSingletonAsSets(2L, 4L, 6L);
+    // Should appear only for first-as match and path-length match
+    AsPath asPath3a = AsPath.ofSingletonAsSets(3L, 5L, 6L);
+    // Should never appear
+    AsPath asPath3b = AsPath.ofSingletonAsSets(3L, 4L, 4L, 6L);
+    // Should always appear
+    AsPath bestAsPath = AsPath.ofSingletonAsSets(3L, 4L, 6L);
+    AsPath asPath3c = bestAsPath;
+    AsPath asPath3d = bestAsPath;
+    Ip nextHop2 = new Ip("2.0.0.0");
+    Ip nextHop3a = new Ip("3.0.0.1");
+    Ip nextHop3b = new Ip("3.0.0.2");
+    Ip nextHop3c = new Ip("3.0.0.3");
+    Ip nextHop3d = new Ip("3.0.0.4");
+
+    /*
+     * Common attributes for all routes
+     */
+    Prefix p = Prefix.ZERO;
+    BgpRoute.Builder b =
+        new BgpRoute.Builder()
+            .setNetwork(p)
+            .setProtocol(RoutingProtocol.BGP)
+            .setOriginType(OriginType.INCOMPLETE);
+
+    BgpRoute bestRoute =
+        b.setAsPath(bestAsPath)
+            .setNextHopIp(new Ip("1.1.1.1"))
+            .setOriginatorIp(new Ip("1.1.1.1"))
+            .setReceivedFromIp(new Ip("1.1.1.1."))
+            .build();
+
+    /*
+     * Instantiate routes
+     */
+    BgpRoute route2 =
+        b.setAsPath(asPath2)
+            .setNextHopIp(nextHop2)
+            .setOriginatorIp(nextHop2)
+            .setReceivedFromIp(nextHop2)
+            .build();
+    BgpRoute route3a =
+        b.setAsPath(asPath3a)
+            .setNextHopIp(nextHop3a)
+            .setOriginatorIp(nextHop3a)
+            .setReceivedFromIp(nextHop3a)
+            .build();
+    BgpRoute route3b =
+        b.setAsPath(asPath3b)
+            .setNextHopIp(nextHop3b)
+            .setOriginatorIp(nextHop3b)
+            .setReceivedFromIp(nextHop3b)
+            .build();
+    BgpRoute route3c =
+        b.setAsPath(asPath3c)
+            .setNextHopIp(nextHop3c)
+            .setOriginatorIp(nextHop3c)
+            .setReceivedFromIp(nextHop3c)
+            .build();
+    BgpRoute route3d =
+        b.setAsPath(asPath3d)
+            .setNextHopIp(nextHop3d)
+            .setOriginatorIp(nextHop3d)
+            .setReceivedFromIp(nextHop3d)
+            .build();
+
+    BgpRib bmr =
+        new BgpRib(
+            null, null, BgpTieBreaker.ARRIVAL_ORDER, null, multipathEquivalentAsPathMatchMode);
+
+    /*
+     * Add routes to multipath RIB.
+     */
+    bmr.mergeRoute(bestRoute);
+    bmr.mergeRoute(route2);
+    bmr.mergeRoute(route3a);
+    bmr.mergeRoute(route3b);
+    bmr.mergeRoute(route3c);
+    bmr.mergeRoute(route3d);
+
+    /*
+     * Initialize the matchers with respect to the output route set
+     */
+    Set<BgpRoute> postMergeRoutes = bmr.getRoutes();
+    Matcher<BgpRoute> present = in(postMergeRoutes);
+    Matcher<BgpRoute> absent = not(present);
+
+    /*
+     * ASSERTIONS:
+     * Only the expected routes for the given match mode should be present at end
+     */
+    assertThat(route2, expectRoute2 ? present : absent);
+    assertThat(route3a, expectRoute3a ? present : absent);
+    assertThat(route3b, expectRoute3b ? present : absent);
+    assertThat(route3c, expectRoute3c ? present : absent);
+    assertThat(route3c, expectRoute3d ? present : absent);
+  }
+
+  @Test
+  public void testBgpAsPathMultipathExactPath() {
+    /*
+     * Only routes with exact AS path match to that of best AS path should appear in RIB post-merge.
+     */
+    testBgpAsPathMultipathHelper(
+        MultipathEquivalentAsPathMatchMode.EXACT_PATH, false, false, false, true, true);
+  }
+
+  @Test
+  public void testBgpAsPathMultipathFirstAs() {
+    /*
+     * Only routes with first-as matching that of best as path should appear in RIB post-merge.
+     */
+    testBgpAsPathMultipathHelper(
+        MultipathEquivalentAsPathMatchMode.FIRST_AS, false, true, false, true, true);
+  }
+
+  @Test
+  public void testBgpAsPathMultipathPathLength() {
+    /*
+     * All routes with same as-path-length as that of best as-path should appear in RIB post-merge.
+     */
+    testBgpAsPathMultipathHelper(
+        MultipathEquivalentAsPathMatchMode.PATH_LENGTH, true, true, false, true, true);
+  }
+
+  @Test
+  public void testBgpCompareOriginType() {
+    BgpRib bbr =
+        new BgpRib(
+            null, null, BgpTieBreaker.ROUTER_ID, 1, MultipathEquivalentAsPathMatchMode.EXACT_PATH);
+    BgpRib bmr =
+        new BgpRib(
+            null,
+            null,
+            BgpTieBreaker.ROUTER_ID,
+            null,
+            MultipathEquivalentAsPathMatchMode.EXACT_PATH);
+
+    Prefix p = Prefix.ZERO;
+    BgpRoute.Builder b = new BgpRoute.Builder().setNetwork(p).setProtocol(RoutingProtocol.IBGP);
+
+    /*
+     *  Initialize with different originator ips, which should not affect comparison of routes with
+     *  different origin type.
+     */
+    Map<OriginType, List<BgpRoute>> routesByOriginType = new LinkedHashMap<>();
+    for (OriginType originType : OriginType.values()) {
+      List<BgpRoute> routes =
+          routesByOriginType.computeIfAbsent(originType, o -> new ArrayList<>());
+      routes.add(
+          b.setOriginatorIp(Ip.ZERO).setReceivedFromIp(Ip.ZERO).setOriginType(originType).build());
+      routes.add(
+          b.setOriginatorIp(Ip.MAX).setReceivedFromIp(Ip.MAX).setOriginType(originType).build());
+    }
+
+    /*
+     * Whenever origin type is different, it should be overriding factor in preference.
+     */
+    for (OriginType o1 : OriginType.values()) {
+      List<BgpRoute> lhsList = routesByOriginType.get(o1);
+      for (OriginType o2 : OriginType.values()) {
+        List<BgpRoute> rhsList = routesByOriginType.get(o2);
+        for (BgpRoute lhs : lhsList) {
+          for (BgpRoute rhs : rhsList) {
+            if (o1.getPreference() > o2.getPreference()) {
+              assertThat(bbr.comparePreference(lhs, rhs), greaterThan(0));
+              assertThat(bmr.comparePreference(lhs, rhs), greaterThan(0));
+            } else if (o1.getPreference() < o2.getPreference()) {
+              assertThat(bbr.comparePreference(lhs, rhs), lessThan(0));
+              assertThat(bmr.comparePreference(lhs, rhs), lessThan(0));
+            }
+          }
+        }
+      }
+    }
+  }
+
+  @Test
+  public void testContainsRoute() {
+
+    BgpRib bbr =
+        new BgpRib(
+            null, null, BgpTieBreaker.ROUTER_ID, 1, MultipathEquivalentAsPathMatchMode.EXACT_PATH);
+    BgpRib bmr =
+        new BgpRib(
+            null,
+            null,
+            BgpTieBreaker.ROUTER_ID,
+            null,
+            MultipathEquivalentAsPathMatchMode.EXACT_PATH);
+    Ip ip1 = new Ip("1.0.0.0");
+    Ip ip2 = new Ip("2.2.0.0");
+    BgpRoute.Builder b1 =
+        new BgpRoute.Builder()
+            .setNextHopIp(Ip.ZERO)
+            .setOriginType(OriginType.INCOMPLETE)
+            .setOriginatorIp(Ip.ZERO)
+            .setProtocol(RoutingProtocol.BGP)
+            .setReceivedFromIp(Ip.ZERO);
+    BgpRoute.Builder b2 =
+        new BgpRoute.Builder()
+            .setNextHopIp(Ip.ZERO)
+            .setOriginType(OriginType.INCOMPLETE)
+            .setOriginatorIp(Ip.MAX)
+            .setProtocol(RoutingProtocol.BGP)
+            .setReceivedFromIp(Ip.ZERO);
+
+    /*
+     * Toss a bunch of different routes in each RIB. In the best-path rib, only lower originatorIp
+     * routes should remain. In the multipath RIB, all routes should remain.
+     */
+    for (int i = 8; i <= Prefix.MAX_PREFIX_LENGTH; i++) {
+      Prefix p = new Prefix(ip1, i);
+      b1.setNetwork(p);
+      b2.setNetwork(p);
+      bbr.mergeRoute(b1.build());
+      bbr.mergeRoute(b2.build());
+      bmr.mergeRoute(b1.build());
+      bmr.mergeRoute(b2.build());
+    }
+    for (int i = 16; i <= Prefix.MAX_PREFIX_LENGTH; i++) {
+      Prefix p = new Prefix(ip2, i);
+      b1.setNetwork(p);
+      b2.setNetwork(p);
+      bbr.mergeRoute(b1.build());
+      bbr.mergeRoute(b2.build());
+      bmr.mergeRoute(b1.build());
+      bmr.mergeRoute(b2.build());
+    }
+    for (int i = 8; i <= Prefix.MAX_PREFIX_LENGTH; i++) {
+      Prefix p = new Prefix(ip1, i);
+      assertTrue(bbr.containsRoute(b1.setNetwork(p).build()));
+      b1.setNetwork(p);
+      b2.setNetwork(p);
+      assertTrue(bbr.containsRoute(b1.build()));
+      assertFalse(bbr.containsRoute(b2.build()));
+      assertTrue(bmr.containsRoute(b1.build()));
+      assertTrue(bmr.containsRoute(b2.build()));
+    }
+    for (int i = 16; i <= Prefix.MAX_PREFIX_LENGTH; i++) {
+      Prefix p = new Prefix(ip2, i);
+      b1.setNetwork(p);
+      b2.setNetwork(p);
+      assertTrue(bbr.containsRoute(b1.build()));
+      assertFalse(bbr.containsRoute(b2.build()));
+      assertTrue(bmr.containsRoute(b1.build()));
+      assertTrue(bmr.containsRoute(b2.build()));
+    }
+  }
+
+  @Test
+  public void testBgpTieBreaker() {
+    // good for both ebgp and ibgp
+    BgpRib bmr =
+        new BgpRib(
+            null,
+            null,
+            BgpTieBreaker.ROUTER_ID,
+            null,
+            MultipathEquivalentAsPathMatchMode.EXACT_PATH);
+    // ebgp
+    BgpRib ebgpBpr =
+        new BgpRib(
+            null, null, BgpTieBreaker.ROUTER_ID, 1, MultipathEquivalentAsPathMatchMode.EXACT_PATH);
+    BgpRoute.Builder ebgpBuilder =
+        new BgpRoute.Builder()
+            .setNetwork(Prefix.ZERO)
+            .setOriginType(OriginType.INCOMPLETE)
+            .setOriginatorIp(Ip.ZERO)
+            .setProtocol(RoutingProtocol.BGP)
+            .setReceivedFromIp(Ip.ZERO);
+    BgpRoute ebgpOlderHigherOriginator =
+        ebgpBuilder.setOriginatorIp(Ip.MAX).setReceivedFromIp(new Ip("1.1.1.1")).build();
+    BgpRoute ebgpNewerHigherOriginator =
+        ebgpBuilder.setOriginatorIp(Ip.MAX).setReceivedFromIp(new Ip("1.1.1.2")).build();
+    BgpRoute ebgpLowerOriginator = ebgpBuilder.setOriginatorIp(Ip.ZERO).build();
+    // ibgp
+    BgpRib ibgpBpr =
+        new BgpRib(
+            null, null, BgpTieBreaker.ROUTER_ID, 1, MultipathEquivalentAsPathMatchMode.EXACT_PATH);
+    BgpRoute.Builder ibgpBuilder =
+        new BgpRoute.Builder()
+            .setNetwork(Prefix.ZERO)
+            .setOriginType(OriginType.INCOMPLETE)
+            .setOriginatorIp(Ip.ZERO)
+            .setProtocol(RoutingProtocol.IBGP)
+            .setReceivedFromIp(Ip.ZERO);
+    BgpRoute ibgpOlderHigherOriginator =
+        ibgpBuilder.setOriginatorIp(Ip.MAX).setReceivedFromIp(new Ip("1.1.1.1")).build();
+    BgpRoute ibgpNewerHigherOriginator =
+        ibgpBuilder.setOriginatorIp(Ip.MAX).setReceivedFromIp(new Ip("1.1.1.2")).build();
+    BgpRoute ibgpLowerOriginator = ibgpBuilder.setOriginatorIp(Ip.ZERO).build();
+
+    ebgpBpr.mergeRoute(ebgpOlderHigherOriginator);
+    ibgpBpr.mergeRoute(ibgpOlderHigherOriginator);
+
+    /*
+     * Given default tie-breaking, and all more important attributes being equivalent:
+     * - When comparing two eBGP adverts, best-path rib prefers older advert.
+     * - If neither is older, or one is iBGP, best-path rib prefers advert with higher router-id.
+     * - Multipath RIB ignores both age and router-id, seeing both adverts as equal.
+     */
+
+    // Test age comparisons first
+    assertThat(
+        ebgpBpr.comparePreference(ebgpNewerHigherOriginator, ebgpOlderHigherOriginator),
+        lessThan(0));
+    assertThat(bmr.comparePreference(ebgpNewerHigherOriginator, ebgpLowerOriginator), equalTo(0));
+    assertThat(
+        ibgpBpr.comparePreference(ibgpNewerHigherOriginator, ibgpLowerOriginator), lessThan(0));
+    assertThat(bmr.comparePreference(ibgpNewerHigherOriginator, ibgpLowerOriginator), equalTo(0));
   }
 }
